requirements.txt
fastapi==0.112.3
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.8.0
httpx>=0.24,<0.26
websockets>=11.0
supabase==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.0.0
redis>=4.0.0
